fix(task): Fix memory leak from `task::run_on_core_nonblocking` * Remove call to `vTaskDelete(nullptr)` from the end of the thread function
`vTaskDelete(nullptr)` was added as part of #353 based on spec. It was erroneously causing the detached thread to leak memory for each call to `run_on_core_nonblocking`. Using esp heap tracing I was able to determine that it did leak memory, while the updated code does not leak memory. Additionally, I used the `TaskMonitor` to ensure that the tasks are properly destroyed and are no longer registered within the system.

* Build and run `task/example` on QtPy ESP32s3 and ensure it works still.
* Build and run modified `task/example` on QtPy ESP32s3 which as freertos runtime stats (for task monitoring) and heap tracing enabled. Dump the heap trace output to ensure the residual memory allocated is not dependent on the number of times `run_on_core_nonblocking` was called. Confirmed that when including `vTaskDelay(nullptr)` the memory allocated (and leaked) is dependent on the number of times it is called.