feat(display): Update to support multiple displays * Update `espp::Display` to store the `flush` and `rotate` functions and to pass `this` as the context for the lvgl `display` and `event` callback
This allows any arbitrary function (including bound class member functions) to be passed for the flush callback and rotate callbacks, instead of requiring that they match the signatures and linkage of the lvgl lower-layer functions. This will enable a follow-up PR which allows the display drivers to no longer be static classes, further enabling multiple displays to be used in a single project.

Build and run `esp-box/example`, `t-dongle-s3/example` to ensure the existing code still works.