feat(icm42670): Add ICM42607 / ICM42670 IMU component * Add new component for IMU * Add IMU to esp-box component and test it in the example
IMUs are nice, and exposing the IMU via the hardware component / bsp (esp-box in this case) is even better. Testing / showcasing the IMU in the bsp example is _even_ better.

Build and run `esp-box/example` on ESP-BOX and ensure the IMU works by looking at the on-screen-display while moving the box around various orientations.