fix(joystick): Update joystick to fix edge case (#411)
* fix(joystick): Update joystick to fix edge case
* Update to ensure we handle the edge case that `center_deadzone_radius == range_deadzone == 0.5f`.

In this case, both of the first conditionals would return false, then magnitude range would be 0.0, leading to div/0 error.

This PR fixes it and ensures that there is not a case for which the magnitude_range will be calculated to be 0.

Build and run `joystick/example` on QtPy ESP32-S3 and ensure new test passes

* update example to explicitly test large / overlapping deadzones