fix(math): Fix range mapper deadband logic (#423)
* Fix issue in range mapper which could occur if the center and range deadzones overlapped
* More explicitly allow uni-directional output mappings when min==center or max==center
* Flesh out test cases for floating point precision issues and values around the deadbands

It was found that the output could be discontinuous if the center deadzone overlapped with one of the range deadzones. The solution was to better handle the case that there are unidirectional output distributions (e.g. where center==min or center==max) and to better handle how the deadbands are calculated.

Build and run `math/example` on QtPy ESP32s3 and ensure the output of `rm4` matches what we expect.